backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.1
Flask-Limiter==3.12
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
bcrypt==4.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
pytest==8.3.5
Werkzeug==3.1.3
